+++ src/EventStore.Core.Tests/ClientOperations/when_starting_a_transaction_expecting_version_any.cs
@@ -9,7 +9,7 @@ using NUnit.Framework;
 
 namespace EventStore.Core.Tests.ClientOperations {
 	[TestFixture(typeof(LogFormat.V2), typeof(string))]
-	[TestFixture(typeof(LogFormat.V3), typeof(long), Ignore = "Explicit transactions are not supported yet by Log V3")]
+	[TestFixture(typeof(LogFormat.V3), typeof(uint), Ignore = "Explicit transactions are not supported yet by Log V3")]
 	public class when_starting_a_transaction_expecting_version_any<TLogFormat, TStreamId> : specification_with_request_manager_integration<TLogFormat, TStreamId> {
 		readonly string _streamId = $"new_test_stream_{Guid.NewGuid()}";
