+++ src/EventStore.Core.Tests/Services/Storage/AllReader/when_reading_from_stream_which_is_disallowed_from_all.cs
@@ -10,8 +10,8 @@ using static EventStore.Core.Messages.TcpClientMessageDto.Filter;
 namespace EventStore.Core.Tests.Services.Storage.AllReader {
 	[TestFixture(typeof(LogFormat.V2), typeof(string), "$persistentsubscription-$all::group-checkpoint")]
 	[TestFixture(typeof(LogFormat.V2), typeof(string), "$persistentsubscription-$all::group-parked")]
-	[TestFixture(typeof(LogFormat.V3), typeof(long), "$persistentsubscription-$all::group-checkpoint")]
-	[TestFixture(typeof(LogFormat.V3), typeof(long), "$persistentsubscription-$all::group-parked")]
+	[TestFixture(typeof(LogFormat.V3), typeof(uint), "$persistentsubscription-$all::group-checkpoint")]
+	[TestFixture(typeof(LogFormat.V3), typeof(uint), "$persistentsubscription-$all::group-parked")]
 	public class when_reading_from_stream_which_is_disallowed_from_all<TLogFormat, TStreamId> : ReadIndexTestScenario<TLogFormat, TStreamId> {
 		private string _stream;
 		public when_reading_from_stream_which_is_disallowed_from_all(string stream) {
