+++ src/EventStore.Projections.Core.Tests/Services/emitted_stream/when_handling_emits_with_previously_written_events.cs
@@ -8,7 +8,7 @@ using NUnit.Framework;
 
 namespace EventStore.Projections.Core.Tests.Services.emitted_stream {
 	[TestFixture(typeof(LogFormat.V2), typeof(string))]
-	[TestFixture(typeof(LogFormat.V3), typeof(long))]
+	[TestFixture(typeof(LogFormat.V3), typeof(uint))]
 	public class when_handling_emits_with_previously_written_events<TLogFormat, TStreamId> : TestFixtureWithExistingEvents<TLogFormat, TStreamId> {
 		private EmittedStream _stream;
 		private TestCheckpointManagerMessageHandler _readyHandler;
