+++ src/EventStore.Projections.Core.Tests/Services/feed_reader/feed_reader.cs
@@ -264,7 +264,7 @@ namespace EventStore.Projections.Core.Tests.Services.feed_reader {
 		}
 
 		[TestFixture(typeof(LogFormat.V2), typeof(string))]
-		[TestFixture(typeof(LogFormat.V3), typeof(long))]
+		[TestFixture(typeof(LogFormat.V3), typeof(uint))]
 		class when_reading_existing_events<TLogFormat, TStreamId> : TestFixtureWithFeedReaderService<TLogFormat, TStreamId> {
 			private QuerySourcesDefinition _querySourcesDefinition;
 			private CheckpointTag _fromPosition;
@@ -303,7 +303,7 @@ namespace EventStore.Projections.Core.Tests.Services.feed_reader {
 		}
 
 		[TestFixture(typeof(LogFormat.V2), typeof(string))]
-		[TestFixture(typeof(LogFormat.V3), typeof(long))]
+		[TestFixture(typeof(LogFormat.V3), typeof(uint))]
 		class when_reading_existing_events_by_parts<TLogFormat, TStreamId> : TestFixtureWithFeedReaderService<TLogFormat, TStreamId> {
 			private QuerySourcesDefinition _querySourcesDefinition;
 			private CheckpointTag _fromPosition;
